Canvas has been abstracted out, a submissions bug still remains

+++ IguideME.Web/Services/Workers/GradePredictorWorker.cs
@@ -54,57 +54,13 @@ namespace IguideME.Web.Services.Workers
             }
         }
 
-        // This version looks at the submissions instead of the average grades and might be useful in the future.
-        // /// <summary>
-        // /// Make predictions for a student
-        // /// </summary>
-        // /// <param name="student">the student to make predictions for.</param>
-        // private void PredictGradesForStudent(User student)
-        // {
-        //     // _logger.LogInformation("Processing student: {ID}", student.UserID);
-        //     List<TileEntrySubmission> submissions = DatabaseManager.Instance.GetCourseSubmissionsForStudent(this._courseID,
-        //                                                                               student.UserID,
-        //                                                                               this._hashCode);
-
-        //     _logger.LogInformation("Submissions {sub}", submissions);
-        //     submissions.ForEach(sub => _logger.LogInformation("eid: {eid}, uid: {uid}", sub.EntryID, sub.UserID));
-        //     List<TileEntry> tileEntries = DatabaseManager.Instance.GetEntries(this._courseID);
-
-        //     double wGrade = 0.0;
-
-        //     foreach (GradePredictionModelParameter modelParameter in this._model.Parameters)
-        //     {
-        //         TileEntrySubmission submission = submissions.Find(sub => sub.EntryID == modelParameter.ParameterID);
-        //         if (submission == null)
-        //         {
-        //             _logger.LogInformation("Student {UserID} has no submission for EntryID {ParamID}", student.UserID, modelParameter.ParameterID);
-        //             continue;
-        //         }
-
-        //         if (!double.TryParse(submission.Grade, out double partialGrade))
-        //         {
-        //             _logger.LogInformation("Error parsing submission grade as double: {Grade}", submission.Grade);
-        //             _logger.LogInformation("Aborting.");
-        //             return;
-        //         }
-
-        //         _logger.LogInformation("grade += {partialGrade} * {Weight}", partialGrade, modelParameter.Weight);
-
-        //         wGrade += partialGrade * modelParameter.Weight;
-        //     }
-
-        //     DatabaseManager.Instance.CreatePredictedGrade(this._courseID,
-        //                                                   student.UserID,
-        //                                                   wGrade);
-        // }
-
         /// Make predictions for a student
         /// </summary>
         /// <param name="student">the student to make predictions for.</param>
         private void PredictGradesForStudent(User student)
         {
 
-            Dictionary<int, List<float>> grades = DatabaseManager.Instance.GetUserGrades(this._courseID,
+            Dictionary<int, List<double>> grades = DatabaseManager.Instance.GetUserGrades(this._courseID,
                                                                                       student.UserID,
                                                                                       this._hashCode);
 
@@ -112,9 +68,9 @@ namespace IguideME.Web.Services.Workers
 
             foreach (GradePredictionModelParameter modelParameter in this._model.Parameters)
             {
-                float partialGrade = 0;
+                double partialGrade = 0;
 
-                if (grades.TryGetValue(modelParameter.ParameterID, out List<float> entryGrades))
+                if (grades.TryGetValue(modelParameter.ParameterID, out List<double> entryGrades))
                 {
                     partialGrade = entryGrades.Average();
                 }
